Telephony: Don't read records from card if not needed
Specifically handle the scenario where there is an RUIM card inserted
but we are in NV mode. In this case, don't read records from the card.
IccCardProxy tells the IccRecord that it will be requiring records
from the card. RuimRecord waits for this indication before fetching
records from the card.

Change-Id: I5f153197c4dc8d4d392be319629718901db104ed